% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflow_doc.R
\name{reflow_doc}
\alias{reflow_doc}
\title{Reflow (reindent to margin) Rmarkdown document}
\usage{
reflow_doc(cast_from = ".")
}
\arguments{
\item{case_from}{Where to cast from}
}
\description{
Reflow (reindent to margin) Rmarkdown document
}
